+ renderowanie do dwóch tekstur + wyświetlanie na podstawie danych z obu tekstur + model jest teraz floatem, bo nie udało się przekonać shaderów do pokachania byte'a + zalążki informacji o różnicach z dokumentacją - info o formacie danych w G-buforze
git-svn-id: svn+ssh://gamma/home/samba/koturj/svn/inz@97 e70bc5d4-e669-4a26-a90f-bfde2a246bde

+++ src/mem/misc/holder.h
@@ -42,7 +42,7 @@ namespace MISC
 		}
 
 		//   * GFX
-		GBUF<uint8_t>      model;
+		GBUF<float>    model;
 
 		//   * COMMON
 		GBUF<float3>   pos;
